+++ src/lib/converter/symbols.ts
@@ -658,6 +658,15 @@ function convertClassOrInterface(
 
     context.finalizeDeclarationReflection(reflection);
 
+    // Convert type parameters early so that we get access them when converting the constructors if any
+    if (instanceType.typeParameters) {
+        reflection.typeParameters = instanceType.typeParameters.map((param) => {
+            const declaration = param.symbol.declarations?.[0];
+            assert(declaration && ts.isTypeParameterDeclaration(declaration));
+            return createTypeParamReflection(declaration, reflectionContext);
+        });
+    }
+
     if (classDeclaration && reflection.kind === ReflectionKind.Class) {
         // Classes can have static props
         const staticType = context.checker.getTypeOfSymbolAtLocation(
@@ -711,15 +720,6 @@ function convertClassOrInterface(
         reflectionContext,
         context.checker.getPropertiesOfType(instanceType),
     );
-
-    // And type arguments
-    if (instanceType.typeParameters) {
-        reflection.typeParameters = instanceType.typeParameters.map((param) => {
-            const declaration = param.symbol.declarations?.[0];
-            assert(declaration && ts.isTypeParameterDeclaration(declaration));
-            return createTypeParamReflection(declaration, reflectionContext);
-        });
-    }
 
     // Interfaces might also have call signatures
     // Classes might too, because of declaration merging