+++ src/test/converter2/behavior/ctorTypeParam.ts
@@ -1,0 +1,10 @@
+/**
+ * @template T class docs
+ */
+export class Generic<T> {
+    /** @template T ctor docs */
+    constructor();
+    constructor(x: string);
+
+    constructor(x?: string) {}
+}