+++ src/test/converter2/issues/gh3031.js
@@ -1,0 +1,7 @@
+/**
+ * Test for Type Parameters on Constructor Signatures being ignored.
+ *
+ * @template T - This is documentation for T
+ */
+export class Test {
+}