+++ src/test/Repository.test.ts
@@ -155,6 +155,16 @@ describe("Repository", function () {
                 repo.getURL(normalizePath(join(project.cwd, "test.js")), 1),
                 "b53cc55bcdd9bc5920787a1d4a4a15fa24123b04/b53cc55b/test.js/1",
             );
+
+            const repo2 = GitRepository.tryCreateRepository(
+                project.cwd,
+                "{gitRevision}/{gitRevision:short}/{path}/{line}",
+                "{branch}", // revision
+                "origin", // remote
+                new TestLogger(),
+            );
+
+            equal(repo2?.gitRevision, "test");
         });
     });
 });