+++ CHANGELOG.md
@@ -13,13 +13,16 @@ title: Changelog
   API: Introduced `typeAnnotation` on `CommentTag`
 - Added `excludePrivateClassFields` option to hide `#private` members while allowing `private` members, #3017.
 - Added support for TypeScript's `@this` tag for JS files which describe `this` parameters, #3026.
-- Remove the `@jsx` tag from the list of additional block tags. TSDoc now directly supports this tag, #3035.
+- API: Re-introduced `relevanceBoost` on `DeclarationReflection` for plugin use, #3036.
 
 ## Bug Fixes
 
 - Fixed conversion of auto-accessor types on properties with the `accessor` keyword, #3019.
 - Improved handling of HTML tags within headers for anchor generation, #3023.
 - Improved support for detecting destructured parameters and renaming them to the name used in the doc comment, #3026.
+- Constructor type parameters will now inherit their class's type parameter descriptions if not otherwise specified, #3031.
+- Fixed compatibility with `@microsoft/tsdoc-config` version 0.18.0, #3035.
+- Custom theme icons will now be used in the "On This Page" sidebar, #3039.
 
 ## v0.28.13 (2025-09-14)